api: Added Common.CostModelParams, Broke testlib
