+++ plutus-ledger-api/src/PlutusLedgerApi/Common.hs
@@ -3,14 +3,15 @@ module PlutusLedgerApi.Common
     ( module PlutusLedgerApi.Common.Eval
     , module PlutusLedgerApi.Common.SerialisedScript
     , module PlutusLedgerApi.Common.Versions
-    , module PlutusLedgerApi.Internal.ParamName
+    , module PlutusLedgerApi.Common.CostModelParams
+    , IsParamName (showParamName)
     ) where
 
 import PlutusLedgerApi.Common.Eval
 import PlutusLedgerApi.Common.SerialisedScript
 import PlutusLedgerApi.Common.Versions
+import PlutusLedgerApi.Common.CostModelParams
 
 import PlutusLedgerApi.Internal.ParamName
     (IsParamName (showParamName)
     )
-
